fix(theme): reapply application night mode when theme changes (#677)
ThemeViewModel.setThemeMode now calls applyPlatformNightMode so LIGHT/DARK/SYSTEM updates UiModeManager live. The helper lives in shared so attachBaseContext and the ViewModel share one path.
