+++ shared/src/commonMain/kotlin/com/devil/phoenixproject/ui/theme/PlatformNightMode.kt
@@ -1,0 +1,4 @@
+package com.devil.phoenixproject.ui.theme
+
+/** Applies LIGHT/DARK/SYSTEM to the Android application night mode. No-op on iOS. */
+expect fun applyPlatformNightMode(themeMode: ThemeMode)
