+++ shared/src/iosMain/kotlin/com/devil/phoenixproject/ui/theme/PlatformNightMode.ios.kt
@@ -1,0 +1,3 @@
+package com.devil.phoenixproject.ui.theme
+
+actual fun applyPlatformNightMode(themeMode: ThemeMode) = Unit
